added password reset

+++ client/src/services/RemoveRecordService.js
@@ -1,7 +1,8 @@
 import Api from '@/services/Api'
+import Store from '@/store/store'
 
 export default {
   delete (formId) {
-    return Api().delete('delete', {params: {formId: formId}})
+    return Api().delete('delete', {params: {formId: formId}, headers: {'Authorization': Store.state.token}})
   }
 }
